+++ src/android/ConnectPlugin.java
@@ -3,8 +3,11 @@ package org.apache.cordova.facebook;
 import android.content.Context;
 import android.content.Intent;
 import android.content.res.Resources;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
 import android.net.Uri;
 import android.os.Bundle;
+import android.util.Base64;
 import android.util.Log;
 import android.webkit.WebView;
 
@@ -21,6 +24,7 @@ import com.facebook.GraphRequest;
 import com.facebook.GraphResponse;
 import com.facebook.HttpMethod;
 import com.facebook.FacebookAuthorizationException;
+import com.facebook.Profile;
 import com.facebook.appevents.AppEventsLogger;
 import com.facebook.applinks.AppLinkData;
 import com.facebook.internal.AttributionIdentifiers;
@@ -30,6 +34,8 @@ import com.facebook.share.Sharer;
 import com.facebook.share.model.GameRequestContent;
 import com.facebook.share.model.ShareHashtag;
 import com.facebook.share.model.ShareLinkContent;
+import com.facebook.share.model.SharePhoto;
+import com.facebook.share.model.SharePhotoContent;
 import com.facebook.share.widget.GameRequestDialog;
 import com.facebook.share.widget.MessageDialog;
 import com.facebook.share.widget.ShareDialog;
@@ -90,9 +96,6 @@ public class ConnectPlugin extends CordovaPlugin {
 
         // augment web view to enable hybrid app events
         enableHybridAppEvents();
-
-        // Set up the activity result callback to this class
-        cordova.setActivityResultCallback(this);
 
         LoginManager.getInstance().registerCallback(callbackManager, new FacebookCallback<LoginResult>() {
             @Override
@@ -264,7 +267,23 @@ public class ConnectPlugin extends CordovaPlugin {
 
     @Override
     public boolean execute(String action, JSONArray args, final CallbackContext callbackContext) throws JSONException {
-        if (action.equals("login")) {
+        if (action.equals("getApplicationId")) {
+            callbackContext.success(FacebookSdk.getApplicationId());
+            return true;
+
+        } else if (action.equals("setApplicationId")) {
+            executeSetApplicationId(args, callbackContext);
+            return true;
+
+        } else if (action.equals("getApplicationName")) {
+            callbackContext.success(FacebookSdk.getApplicationName());
+            return true;
+
+        } else if (action.equals("setApplicationName")) {
+            executeSetApplicationName(args, callbackContext);
+            return true;
+
+        } else if (action.equals("login")) {
             executeLogin(args, callbackContext);
             return true;
 
@@ -311,8 +330,20 @@ public class ConnectPlugin extends CordovaPlugin {
             return true;
 
         } else if(action.equals("setAdvertiserIDCollectionEnabled")) {
-          executeSetAdvertiserIDCollectionEnabled(args, callbackContext);
-          return true;
+            executeSetAdvertiserIDCollectionEnabled(args, callbackContext);
+            return true;
+
+        } else if(action.equals("setDataProcessingOptions")) {
+            executeSetDataProcessingOptions(args, callbackContext);
+            return true;
+
+        } else if (action.equals("setUserData")) {
+            executeSetUserData(args, callbackContext);
+            return true;
+
+        } else if (action.equals("clearUserData")) {
+            executeClearUserData(args, callbackContext);
+            return true;
 
         } else if (action.equals("logEvent")) {
             executeLogEvent(args, callbackContext);
@@ -326,6 +357,10 @@ public class ConnectPlugin extends CordovaPlugin {
             executeDialog(args, callbackContext);
             return true;
 
+        } else if (action.equals("getCurrentProfile")) {
+            executeGetCurrentProfile(args, callbackContext);
+
+            return true;
         } else if (action.equals("graphApi")) {
             executeGraph(args, callbackContext);
 
@@ -345,6 +380,38 @@ public class ConnectPlugin extends CordovaPlugin {
             return true;
         }
         return false;
+    }
+
+    private void executeSetApplicationId(JSONArray args, CallbackContext callbackContext) {
+        if (args.length() == 0) {
+            // Not enough parameters
+            callbackContext.error("Invalid arguments");
+            return;
+        }
+
+        try {
+            String appId = args.getString(0);
+            FacebookSdk.setApplicationId(appId);
+            callbackContext.success();
+        } catch (JSONException e) {
+            callbackContext.error("Error setting application ID");
+        }
+    }
+
+    private void executeSetApplicationName(JSONArray args, CallbackContext callbackContext) {
+        if (args.length() == 0) {
+            // Not enough parameters
+            callbackContext.error("Invalid arguments");
+            return;
+        }
+
+        try {
+            String appName = args.getString(0);
+            FacebookSdk.setApplicationName(appName);
+            callbackContext.success();
+        } catch (JSONException e) {
+            callbackContext.error("Error setting application name");
+        }
     }
 
     private void executeGetDeferredApplink(JSONArray args,
@@ -444,7 +511,7 @@ public class ConnectPlugin extends CordovaPlugin {
             gameRequestDialog.show(builder.build());
 
         } else if (method.equalsIgnoreCase("share") || method.equalsIgnoreCase("feed")) {
-            if (!ShareDialog.canShow(ShareLinkContent.class)) {
+            if ((params.containsKey("photo_image") && !ShareDialog.canShow(SharePhotoContent.class)) || (!params.containsKey("photo_image") && !ShareDialog.canShow(ShareLinkContent.class))) {
                 callbackContext.error("Cannot show dialog");
                 return;
             }
@@ -453,10 +520,15 @@ public class ConnectPlugin extends CordovaPlugin {
             pr.setKeepCallback(true);
             showDialogContext.sendPluginResult(pr);
 
-            ShareLinkContent content = buildContent(params);
             // Set up the activity result callback to this class
             cordova.setActivityResultCallback(this);
-            shareDialog.show(content);
+            if (params.containsKey("photo_image")) {
+                SharePhotoContent content = buildPhotoContent(params);
+                shareDialog.show(content);
+            } else {
+                ShareLinkContent content = buildLinkContent(params);
+                shareDialog.show(content);
+            }
 
         } else if (method.equalsIgnoreCase("send")) {
             if (!MessageDialog.canShow(ShareLinkContent.class)) {
@@ -476,6 +548,14 @@ public class ConnectPlugin extends CordovaPlugin {
 
         } else {
             callbackContext.error("Unsupported dialog method.");
+        }
+    }
+
+    private void executeGetCurrentProfile(JSONArray args, CallbackContext callbackContext) {
+        if (Profile.getCurrentProfile() == null) {
+            callbackContext.error("No current profile.");
+        } else {
+            callbackContext.success(getProfile());
         }
     }
 
@@ -543,6 +623,65 @@ public class ConnectPlugin extends CordovaPlugin {
     private void executeSetAdvertiserIDCollectionEnabled(JSONArray args, CallbackContext callbackContext) {
         boolean enabled = args.optBoolean(0);
         FacebookSdk.setAdvertiserIDCollectionEnabled(enabled);
+        callbackContext.success();
+    }
+
+    private void executeSetDataProcessingOptions(JSONArray args, CallbackContext callbackContext) throws JSONException {
+        if (args.length() == 0) {
+            // Not enough parameters
+            callbackContext.error("Invalid arguments");
+            return;
+        }
+
+        JSONArray arr = args.getJSONArray(0);
+        String[] options = new String[arr.length()];
+        for (int i = 0; i < arr.length(); i++) {
+            options[i + 1] = arr.getString(i);
+        }
+
+        if (args.length() == 1) {
+            FacebookSdk.setDataProcessingOptions(options);
+        } else {
+            String country = args.getString(1);
+            String state = args.getString(2);
+            FacebookSdk.setDataProcessingOptions(options);
+        }
+        callbackContext.success();
+    }
+
+    private void executeSetUserData(JSONArray args, CallbackContext callbackContext) throws JSONException {
+        if (args.length() == 0) {
+            // Not enough parameters
+            callbackContext.error("Invalid arguments");
+            return;
+        }
+
+        Map<String, String> params = new HashMap<String, String>();
+        JSONObject parameters;
+
+        try {
+            parameters = args.getJSONObject(0);
+        } catch (JSONException e) {
+            callbackContext.error("userData must be an object");
+            return;
+        }
+
+        Iterator<String> iter = parameters.keys();
+        while (iter.hasNext()) {
+            String key = iter.next();
+            try {
+                params.put(key, parameters.getString(key));
+            } catch (JSONException e) {
+                Log.w(TAG, "Non-string parameter provided to setUserData discarded");
+            }
+        }
+
+        logger.setUserData(params.get("em"), params.get("fn"), params.get("ln"), params.get("ph"), params.get("db"), params.get("ge"), params.get("ct"), params.get("st"), params.get("zp"), params.get("cn"));
+        callbackContext.success();
+    }
+
+    private void executeClearUserData(JSONArray args, CallbackContext callbackContext) {
+        logger.clearUserData();
         callbackContext.success();
     }
 
@@ -733,7 +872,27 @@ public class ConnectPlugin extends CordovaPlugin {
         }
     }
 
-    private ShareLinkContent buildContent(Map<String, String> paramBundle) {
+    private SharePhotoContent buildPhotoContent(Map<String, String> paramBundle) {
+        SharePhoto.Builder photoBuilder = new SharePhoto.Builder();
+        if (!(paramBundle.get("photo_image") instanceof String)) {
+            Log.d(TAG, "photo_image must be a string");
+        } else {
+            try {
+                byte[] photoImageData = Base64.decode(paramBundle.get("photo_image"), Base64.DEFAULT);
+                Bitmap image = BitmapFactory.decodeByteArray(photoImageData, 0, photoImageData.length); 
+                photoBuilder.setBitmap(image).setUserGenerated(true);
+            } catch (Exception e) {
+                Log.d(TAG, "photo_image cannot be decoded");
+            }
+        }
+        SharePhoto photo = photoBuilder.build();
+        SharePhotoContent.Builder photoContentBuilder = new SharePhotoContent.Builder();
+        photoContentBuilder.addPhoto(photo);
+
+        return photoContentBuilder.build();
+    }
+
+    private ShareLinkContent buildLinkContent(Map<String, String> paramBundle) {
         ShareLinkContent.Builder builder = new ShareLinkContent.Builder();
         if (paramBundle.containsKey("href"))
             builder.setContentUrl(Uri.parse(paramBundle.get("href")));
@@ -834,15 +993,15 @@ public class ConnectPlugin extends CordovaPlugin {
         String response;
         final AccessToken accessToken = AccessToken.getCurrentAccessToken();
         if (hasAccessToken()) {
+            long dataAccessExpirationTimeInterval = accessToken.getDataAccessExpirationTime().getTime() / 1000L;
             Date today = new Date();
             long expiresTimeInterval = (accessToken.getExpires().getTime() - today.getTime()) / 1000L;
             response = "{"
                 + "\"status\": \"connected\","
                 + "\"authResponse\": {"
                 + "\"accessToken\": \"" + accessToken.getToken() + "\","
+                + "\"data_access_expiration_time\": \"" + Math.max(dataAccessExpirationTimeInterval, 0) + "\","
                 + "\"expiresIn\": \"" + Math.max(expiresTimeInterval, 0) + "\","
-                + "\"session_key\": true,"
-                + "\"sig\": \"...\","
                 + "\"userID\": \"" + accessToken.getUserId() + "\""
                 + "}"
                 + "}";
@@ -906,6 +1065,26 @@ public class ConnectPlugin extends CordovaPlugin {
 
         response += "\"errorMessage\": \"" + message + "\"}";
 
+        try {
+            return new JSONObject(response);
+        } catch (JSONException e) {
+            e.printStackTrace();
+        }
+        return new JSONObject();
+    }
+    
+    public JSONObject getProfile() {
+        String response;
+        final Profile profile = Profile.getCurrentProfile();
+        if (profile == null) {
+            response = "{}";
+        } else {
+            response = "{"
+                + "\"userID\": \"" + profile.getId() + "\","
+                + "\"firstName\": \"" + profile.getFirstName() + "\","
+                + "\"lastName\": \"" + profile.getLastName() + "\""
+                + "}";
+        }
         try {
             return new JSONObject(response);
         } catch (JSONException e) {